+++ app/project/bootstrappers/databases/Redis.php
@@ -16,7 +16,7 @@ class Redis extends Bootstrapper implements ILazyBootstrapper
     /**
      * {@inheritdoc}
      */
-    public function getBoundClasses()
+    public function getBindings()
     {
         return [IRedis::class];
     }