+++ app/project/bootstrappers/databases/SQL.php
@@ -16,7 +16,7 @@ class SQL extends Bootstrapper implements ILazyBootstrapper
     /**
      * {@inheritdoc}
      */
-    public function getBoundClasses()
+    public function getBindings()
     {
         return [ConnectionPool::class];
     }